dispersion_relation.fit_mode: modify initial guesses for omega_0
Now they are always in the interior of the domain.

This is to work around a problem when the positivity bound on the
amplitude of the Lorentzians is removed: curve_fit would get stuck
in a local optimum, whereby the Lorentzian would be used to fit
the continuum near the boundary.