estimate_sigma: change method for sigma estimation
We now use the larger among an omega-dependent estimate (where the error is estimated using boxcar smoothing of half-width 1 grid point) and a global omega-independent estimate.

The problem with using the smoothing-based estimate alone is that it sometimes becomes very small, leading to particular individual grid points getting overemphasized in the fit.